[razgrebator-client] imid to long

+++ src/main/java/ua/ardas/esputnik/services/clients/razgrebator/RazgrebatorClient.java
@@ -28,7 +28,7 @@ public class RazgrebatorClient {
         resultsQueue = new DefaultRedisList<>(RedisNames.RESULTS_UNSUBSCRIBED, redisTemplate);
     }
 
-    public void processSpam(int orgId, int contactId, Integer messageInstanceId, Integer imId, String feedback) throws InterruptedException {
+    public void processSpam(int orgId, int contactId, Integer messageInstanceId, Long imId, String feedback) throws InterruptedException {
         ComplaintInfo ir = new ComplaintInfo();
         ir.setMessageInstanceId(messageInstanceId);
         ir.setSpam(true);
@@ -62,7 +62,7 @@ public class RazgrebatorClient {
         processUnsubscribed(iid, null);
     }
 
-    public void processUnsubscribed(int orgId, int contactId, Integer messageInstanceId, Integer imId, String feedback) throws InterruptedException {
+    public void processUnsubscribed(int orgId, int contactId, Integer messageInstanceId, Long imId, String feedback) throws InterruptedException {
         ComplaintInfo ir = new ComplaintInfo();
         ir.setMessageInstanceId(messageInstanceId);
         ir.setContactId(contactId);
